+++ cogs/utility.py
@@ -11,15 +11,15 @@ from itertools import takewhile, zip_longest
 from json import JSONDecodeError, loads
 from subprocess import PIPE
 from textwrap import indent
-from types import SimpleNamespace
 from typing import Union
 
 import discord
-from aiohttp import ClientResponseError
 from discord.enums import ActivityType, Status
 from discord.ext import commands, tasks
 from discord.ext.commands.view import StringView
-from pkg_resources import parse_version
+
+from aiohttp import ClientResponseError
+from packaging.version import Version
 
 from core import checks, utils
 from core.changelog import Changelog
@@ -30,7 +30,7 @@ from core.models import (
     UnseenFormatter,
     getLogger,
 )
-from core.utils import trigger_typing, truncate
+from core.utils import trigger_typing, truncate, DummyParam
 from core.paginator import EmbedPaginatorSession, MessagePaginatorSession
 
 
@@ -142,7 +142,7 @@ class ModmailHelpCommand(commands.HelpCommand):
             perm_level = "NONE"
 
         embed = discord.Embed(
-            title=f"`{self.get_command_signature(topic)}`",
+            title=f"`{self.get_command_signature(topic).strip()}`",
             color=self.context.bot.main_color,
             description=self.process_help_msg(topic.help),
         )
@@ -342,9 +342,9 @@ class Utility(commands.Cog):
         latest = changelog.latest_version
 
         if self.bot.version.is_prerelease:
-            stable = next(filter(lambda v: not parse_version(v.version).is_prerelease, changelog.versions))
+            stable = next(filter(lambda v: not Version(v.version).is_prerelease, changelog.versions))
             footer = f"You are on the prerelease version • the latest version is v{stable.version}."
-        elif self.bot.version < parse_version(latest.version):
+        elif self.bot.version < Version(latest.version):
             footer = f"A newer version is available v{latest.version}."
         else:
             footer = "You are up to date with the latest version."
@@ -352,7 +352,7 @@ class Utility(commands.Cog):
         embed.add_field(
             name="Want Modmail in Your Server?",
             value="Follow the installation guide on [GitHub](https://github.com/modmail-dev/modmail/) "
-            "and join our [Discord server](https://discord.gg/F34cRU8)!",
+            "and join our [Discord server](https://discord.gg/cnUpwrnpYb)!",
             inline=False,
         )
 
@@ -401,13 +401,7 @@ class Utility(commands.Cog):
     async def debug(self, ctx):
         """Shows the recent application logs of the bot."""
 
-        log_file_name = self.bot.token.split(".")[0]
-
-        with open(
-            os.path.join(os.path.dirname(os.path.abspath(__file__)), f"../temp/{log_file_name}.log"),
-            "r+",
-            encoding="utf-8",
-        ) as f:
+        with open(self.bot.log_file_path, "r+", encoding="utf-8") as f:
             logs = f.read().strip()
 
         if not logs:
@@ -416,7 +410,7 @@ class Utility(commands.Cog):
                 title="Debug Logs:",
                 description="You don't have any logs at the moment.",
             )
-            embed.set_footer(text="Go to Heroku to see your logs.")
+            embed.set_footer(text="Go to your console to see your logs.")
             return await ctx.send(embed=embed)
 
         messages = []
@@ -433,7 +427,7 @@ class Utility(commands.Cog):
                     msg = "```Haskell\n"
             msg += line
             if len(msg) + 3 > 2000:
-                msg = msg[:1993] + "[...]```"
+                msg = msg[:1992] + "[...]```"
                 messages.append(msg)
                 msg = "```Haskell\n"
 
@@ -455,12 +449,8 @@ class Utility(commands.Cog):
         """Posts application-logs to Hastebin."""
 
         haste_url = os.environ.get("HASTE_URL", "https://hastebin.cc")
-        log_file_name = self.bot.token.split(".")[0]
 
-        with open(
-            os.path.join(os.path.dirname(os.path.abspath(__file__)), f"../temp/{log_file_name}.log"),
-            "rb+",
-        ) as f:
+        with open(self.bot.log_file_path, "rb+") as f:
             logs = BytesIO(f.read().strip())
 
         try:
@@ -482,7 +472,7 @@ class Utility(commands.Cog):
                 color=self.bot.main_color,
                 description="Something's wrong. We're unable to upload your logs to hastebin.",
             )
-            embed.set_footer(text="Go to Heroku to see your logs.")
+            embed.set_footer(text="Go to your console to see your logs.")
         await ctx.send(embed=embed)
 
     @debug.command(name="clear", aliases=["wipe"])
@@ -491,12 +481,7 @@ class Utility(commands.Cog):
     async def debug_clear(self, ctx):
         """Clears the locally cached logs."""
 
-        log_file_name = self.bot.token.split(".")[0]
-
-        with open(
-            os.path.join(os.path.dirname(os.path.abspath(__file__)), f"../temp/{log_file_name}.log"),
-            "w",
-        ):
+        with open(self.bot.log_file_path, "w"):
             pass
         await ctx.send(
             embed=discord.Embed(color=self.bot.main_color, description="Cached logs are now cleared.")
@@ -537,12 +522,12 @@ class Utility(commands.Cog):
             return await ctx.send(embed=embed)
 
         if not message:
-            raise commands.MissingRequiredArgument(SimpleNamespace(name="message"))
+            raise commands.MissingRequiredArgument(DummyParam("message"))
 
         try:
             activity_type = ActivityType[activity_type]
         except KeyError:
-            raise commands.MissingRequiredArgument(SimpleNamespace(name="activity"))
+            raise commands.MissingRequiredArgument(DummyParam("activity"))
 
         activity, _ = await self.set_presence(activity_type=activity_type, activity_message=message)
 
@@ -587,7 +572,7 @@ class Utility(commands.Cog):
         try:
             status = Status[status_type]
         except KeyError:
-            raise commands.MissingRequiredArgument(SimpleNamespace(name="status"))
+            raise commands.MissingRequiredArgument(DummyParam("status"))
 
         _, status = await self.set_presence(status=status)
 
@@ -599,7 +584,6 @@ class Utility(commands.Cog):
         return await ctx.send(embed=embed)
 
     async def set_presence(self, *, status=None, activity_type=None, activity_message=None):
-
         if status is None:
             status = self.bot.config.get("status")
 
@@ -1020,7 +1004,7 @@ class Utility(commands.Cog):
                 color=self.bot.error_color, description="You dont have any aliases at the moment."
             )
             embed.set_footer(text=f'Do "{self.bot.prefix}help alias" for more commands.')
-            embed.set_author(name="Aliases", icon_url=self.bot.get_guild_icon(guild=ctx.guild))
+            embed.set_author(name="Aliases", icon_url=self.bot.get_guild_icon(guild=ctx.guild, size=128))
             return await ctx.send(embed=embed)
 
         embeds = []
@@ -1028,7 +1012,9 @@ class Utility(commands.Cog):
         for i, names in enumerate(zip_longest(*(iter(sorted(self.bot.aliases)),) * 15)):
             description = utils.format_description(i, names)
             embed = discord.Embed(color=self.bot.main_color, description=description)
-            embed.set_author(name="Command Aliases", icon_url=self.bot.get_guild_icon(guild=ctx.guild))
+            embed.set_author(
+                name="Command Aliases", icon_url=self.bot.get_guild_icon(guild=ctx.guild, size=128)
+            )
             embeds.append(embed)
 
         session = EmbedPaginatorSession(ctx, *embeds)
@@ -1115,7 +1101,7 @@ class Utility(commands.Cog):
         await self.bot.config.update()
         return embed
 
-    @alias.command(name="add")
+    @alias.command(name="add", aliases=["create", "make"])
     @checks.has_permissions(PermissionLevel.MODERATOR)
     async def alias_add(self, ctx, name: str.lower, *, value):
         """
@@ -1612,7 +1598,8 @@ class Utility(commands.Cog):
                             )
                             embed = discord.Embed(color=self.bot.main_color, description=description)
                             embed.set_author(
-                                name="Permission Overrides", icon_url=self.bot.get_guild_icon(guild=ctx.guild)
+                                name="Permission Overrides",
+                                icon_url=self.bot.get_guild_icon(guild=ctx.guild, size=128),
                             )
                             embeds.append(embed)
 
@@ -1944,7 +1931,7 @@ class Utility(commands.Cog):
             "(https://github.com/modmail-dev/modmail/blob/master/bot.py#L1)"
         )
 
-        if self.bot.version >= parse_version(latest.version) and flag.lower() != "force":
+        if self.bot.version >= Version(latest.version) and flag.lower() != "force":
             embed = discord.Embed(title="Already up to date", description=desc, color=self.bot.main_color)
 
             data = await self.bot.api.get_user_info()