+++ core/config.py
@@ -21,7 +21,6 @@ load_dotenv()
 
 
 class ConfigManager:
-
     public_keys = {
         # activity
         "twitch_url": "https://www.twitch.tv/discordmodmail/",
@@ -37,6 +36,7 @@ class ConfigManager:
         "account_age": isodate.Duration(),
         "guild_age": isodate.Duration(),
         "thread_cooldown": isodate.Duration(),
+        "log_expiration": isodate.Duration(),
         "reply_without_command": False,
         "anon_reply_without_command": False,
         "plain_reply_without_command": False,
@@ -123,7 +123,7 @@ class ConfigManager:
         # confirm thread creation
         "confirm_thread_creation": False,
         "confirm_thread_creation_title": "Confirm thread creation",
-        "confirm_thread_response": "React to confirm thread creation which will directly contact the moderators",
+        "confirm_thread_response": "Click the button to confirm thread creation which will directly contact the moderators.",
         "confirm_thread_creation_accept": "\N{WHITE HEAVY CHECK MARK}",
         "confirm_thread_creation_deny": "\N{NO ENTRY SIGN}",
         # regex
@@ -178,13 +178,16 @@ class ConfigManager:
         "disable_updates": False,
         # Logging
         "log_level": "INFO",
+        "stream_log_format": "plain",
+        "file_log_format": "plain",
+        "discord_log_level": "INFO",
         # data collection
         "data_collection": True,
     }
 
     colors = {"mod_color", "recipient_color", "main_color", "error_color"}
 
-    time_deltas = {"account_age", "guild_age", "thread_auto_close", "thread_cooldown"}
+    time_deltas = {"account_age", "guild_age", "thread_auto_close", "thread_cooldown", "log_expiration"}
 
     booleans = {
         "use_user_id_channel_name",