+++ core/thread.py
@@ -29,6 +29,10 @@ from core.utils import (
     get_top_role,
     create_thread_channel,
     get_joint_id,
+    AcceptButton,
+    DenyButton,
+    ConfirmThreadCreationView,
+    DummyParam,
 )
 
 logger = getLogger(__name__)
@@ -228,7 +232,9 @@ class Thread:
             else:
                 footer = self.bot.config["thread_creation_footer"]
 
-            embed.set_footer(text=footer, icon_url=self.bot.get_guild_icon(guild=self.bot.modmail_guild))
+            embed.set_footer(
+                text=footer, icon_url=self.bot.get_guild_icon(guild=self.bot.modmail_guild, size=128)
+            )
             embed.title = self.bot.config["thread_creation_title"]
 
             if creator is None or creator == recipient:
@@ -521,7 +527,7 @@ class Thread:
 
         embed.description = message
         footer = self.bot.config["thread_close_footer"]
-        embed.set_footer(text=footer, icon_url=self.bot.get_guild_icon(guild=self.bot.guild))
+        embed.set_footer(text=footer, icon_url=self.bot.get_guild_icon(guild=self.bot.guild, size=128))
 
         if not silent:
             for user in self.recipients:
@@ -716,7 +722,6 @@ class Thread:
     async def find_linked_message_from_dm(
         self, message, either_direction=False, get_thread_channel=False
     ) -> typing.List[discord.Message]:
-
         joint_id = None
         if either_direction:
             joint_id = get_joint_id(message)
@@ -795,8 +800,8 @@ class Thread:
     async def note(
         self, message: discord.Message, persistent=False, thread_creation=False
     ) -> discord.Message:
-        if not message.content and not message.attachments:
-            raise MissingRequiredArgument(SimpleNamespace(name="msg"))
+        if not message.content and not message.attachments and not message.stickers:
+            raise MissingRequiredArgument(DummyParam("msg"))
 
         msg = await self.send(
             message,
@@ -816,8 +821,8 @@ class Thread:
         self, message: discord.Message, anonymous: bool = False, plain: bool = False
     ) -> typing.Tuple[typing.List[discord.Message], discord.Message]:
         """Returns List[user_dm_msg] and thread_channel_msg"""
-        if not message.content and not message.attachments:
-            raise MissingRequiredArgument(SimpleNamespace(name="msg"))
+        if not message.content and not message.attachments and not message.stickers:
+            raise MissingRequiredArgument(DummyParam("msg"))
         if not any(g.get_member(self.id) for g in self.bot.guilds):
             return await message.channel.send(
                 embed=discord.Embed(
@@ -909,7 +914,6 @@ class Thread:
         persistent_note: bool = False,
         thread_creation: bool = False,
     ) -> None:
-
         if not note and from_mod:
             self.bot.loop.create_task(self._restart_close_timer())  # Start or restart thread auto close
 
@@ -957,7 +961,7 @@ class Thread:
                     name = tag
                 avatar_url = self.bot.config["anon_avatar_url"]
                 if avatar_url is None:
-                    avatar_url = self.bot.get_guild_icon(guild=self.bot.guild)
+                    avatar_url = self.bot.get_guild_icon(guild=self.bot.guild, size=128)
                 embed.set_author(
                     name=name,
                     icon_url=avatar_url,
@@ -1016,8 +1020,14 @@ class Thread:
                 return stream.read()
 
         for i in message.stickers:
-            if i.format in (discord.StickerFormatType.png, discord.StickerFormatType.apng):
-                images.append((i.url, i.name, True))
+            if i.format in (
+                discord.StickerFormatType.png,
+                discord.StickerFormatType.apng,
+                discord.StickerFormatType.gif,
+            ):
+                images.append(
+                    (f"https://media.discordapp.net/stickers/{i.id}.{i.format.file_extension}", i.name, True)
+                )
             elif i.format == discord.StickerFormatType.lottie:
                 # save the json lottie representation
                 try:
@@ -1416,30 +1426,19 @@ class ThreadManager:
                 destination = recipient
             else:
                 destination = message.channel
+            view = ConfirmThreadCreationView()
+            view.add_item(AcceptButton(self.bot.config["confirm_thread_creation_accept"]))
+            view.add_item(DenyButton(self.bot.config["confirm_thread_creation_deny"]))
             confirm = await destination.send(
                 embed=discord.Embed(
                     title=self.bot.config["confirm_thread_creation_title"],
                     description=self.bot.config["confirm_thread_response"],
                     color=self.bot.main_color,
-                )
+                ),
+                view=view,
             )
-            accept_emoji = self.bot.config["confirm_thread_creation_accept"]
-            deny_emoji = self.bot.config["confirm_thread_creation_deny"]
-            emojis = [accept_emoji, deny_emoji]
-            for emoji in emojis:
-                await confirm.add_reaction(emoji)
-                await asyncio.sleep(0.2)
-
-            try:
-                r, _ = await self.bot.wait_for(
-                    "reaction_add",
-                    check=lambda r, u: u.id == recipient.id
-                    and r.message.id == confirm.id
-                    and r.message.channel.id == confirm.channel.id
-                    and str(r.emoji) in (accept_emoji, deny_emoji),
-                    timeout=20,
-                )
-            except asyncio.TimeoutError:
+            await view.wait()
+            if view.value is None:
                 thread.cancelled = True
                 self.bot.loop.create_task(
                     destination.send(
@@ -1450,23 +1449,16 @@ class ThreadManager:
                         )
                     )
                 )
-            else:
-                if str(r.emoji) == deny_emoji:
-                    thread.cancelled = True
-                    self.bot.loop.create_task(
-                        destination.send(
-                            embed=discord.Embed(
-                                title=self.bot.config["thread_cancelled"], color=self.bot.error_color
-                            )
+                await confirm.edit(view=None)
+            if view.value is False:
+                thread.cancelled = True
+                self.bot.loop.create_task(
+                    destination.send(
+                        embed=discord.Embed(
+                            title=self.bot.config["thread_cancelled"], color=self.bot.error_color
                         )
                     )
-
-            async def remove_reactions():
-                for emoji in emojis:
-                    await confirm.remove_reaction(emoji, self.bot.user)
-                    await asyncio.sleep(0.2)
-
-            self.bot.loop.create_task(remove_reactions())
+                )
             if thread.cancelled:
                 del self.cache[recipient.id]
                 return thread